# schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: 304CEM – Public Library A Book Lending Web App
  description: An API for A Book Lending Web App
  contact:
    name: Mantas Mikalauskas
    email: mikalaum@coventry.ac.uk
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://ecology-fiction-3000.codio-box.uk/api/v1"
    description: Development server
...
paths:
...
  /books:
    summary: Represents a book
    description: >
      This resource represents a book in the system.
      Each book is identified by a numeric `id`.
    post:
      tags:
        - Books
      description: Add a new book.
      requestBody:
        description: All new book data.
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: ./book.json#/definitions/book
      responses:
        "201":
          description: Confirmation of created book and a link to it
          content:
            application/json:
              schema:
                $ref: ./book.json#/definitions/bookCreated
    get:
      tags:
        - Books
      description: Gets all books
      parameters:
      - in: path
        name: userID
        schema:
          type: integer
        required: true
        description: Numeric id of the user
      - in: query
        name: offset
        schema:
          type: integer
        description: The number of items to skip before starting to collect the result set
      - in: query
        name: limit
        schema:
          type: integer
        description: The numbers of items to return
      - in: query
        name: order
        schema:
          type: string
        description: Field to order results by
      - in: query
        name: direction
        schema:
          type: string
        description: ASC or DESC
      responses:
        "200":
          description: An array of books with pagination
          content:
            application/json:
              schema:
                $ref: ./book.json#/definitions/bookList
  /books/{bookID}:
    parameters:
    - in: path
      name: bookID
      schema:
        type: integer
      required: true
      description: Numeric id of the book
    get:
      tags:
        - Books
      summary: Gets a book by ID
      responses:
        "200":
          description: A book object
          content:
            application/json:
              schema:
                $ref: ./book.json#/definitions/book
        "404":
          description: Not found
    put:
      tags:
        - Books
      summary: Updates a book by ID
      requestBody:
        description: All book update data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./book.json#/definitions/book
      responses:
        "201":
          description: Book updated
          content:
            application/json:
              schema:
                $ref: ./book.json#/definitions/bookUpdated
        "404":
          description: Not found
        "401":
          description: User must be authenticated
        "403":
          description: User must be the owner of the book
    delete:
      tags:
        - Books
      summary: Deletes a book by ID
      responses:
        "201":
          description: Book deleted
          content:
            application/json:
              schema:
                $ref: ./book.json#/definitions/bookDeleted
        "404":
          description: Not found
        "401":
          description: User must be authenticated
        "403":
          description: User must be the owner of the book
...
  /search/books:
    parameters:
    - in: query
      name: q
      schema:
        type: string
      description: A search term.
    - in: query
      name: userID
      schema:
        type: number
      description: (Optional) Search for user books. Must be authenticated 
    - in: query
      name: offset
      schema:
        type: integer
      description: The number of items to skip before starting to collect the result set
    - in: query
      name: limit
      schema:
        type: integer
      description: The numbers of items to return
    - in: query
      name: order
      schema:
        type: string
      description: Field to order results by
    - in: query
      name: direction
      schema:
        type: string
      description: ASC or DESC
    get:
      tags:
        - Search
      summary: Searches for books by title, author or ISBN
      description: >
        Looks for only user books if userID is
        supplied as a query parameter. Searches all books otherwise
      responses:
        "200":
          description: A list of books
          content:
            application/json:
              schema:
                $ref: ./book.json#/definitions/bookList
        "401":
          description: User must be authenticated